requirement.txt
django==4.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.0
django-filter==23.5
psycopg2-binary==2.9.9  # For PostgreSQL
pillow==10.1.0  # For image handling
django-environ
geopy==2.4.0
celery==5.3.4  # For async tasks
redis==5.0.1
django-timezone-field==6.1.0
phonenumbers==8.13.0
razorpay==1.4.1
requests==2.31.0
gunicorn==21.2.0
whitenoise==6.6.0
python-decouple
GDAL==3.8.4
django-phonenumber-field[phonenumbers]
